+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/GetDataStatusResponse.cs
@@ -26,7 +26,7 @@ namespace CodatSyncExpenses.Models.Operations
         /// <summary>
         /// OK
         /// </summary>
-        public Dictionary<string, DataStatus>? DataStatusResponse { get; set; }
+        public Dictionary<string, DataStatus>? DataStatuses { get; set; }
 
         /// <summary>
         /// Your API request was not properly authorized.